Move `clearSession` function to utils
